Fixed typing issues

+++ tests/functional/AbstractBigBlueButtonFunctionalTest.php
@@ -197,7 +197,7 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
         $params = $this->generateJoinMeetingParams();
         $joinMeetingParams = new JoinMeetingParameters($result->getMeetingId(), $params['fullName'], $params['role']);
         $joinMeetingParams->setRedirect(false);
-        $joinMeetingParams->setCreateTime(sprintf('%.0f', $creationTime));
+        $joinMeetingParams->setCreateTime((int) sprintf('%.0f', $creationTime));
 
         $joinMeeting = $this->bbb->joinMeeting($joinMeetingParams);
         $this->assertEquals('SUCCESS', $joinMeeting->getReturnCode(), 'Join meeting');
